+++ src/mlpack/methods/amf/termination_policies/validation_rmse_termination.hpp
@@ -64,7 +64,7 @@ class ValidationRMSETermination
     test_points.zeros(num_test_points, 3);
 
     // fill validation set matrix with random chosen entries
-    for (size_t i = 0; i < num_test_points; i++)
+    for (size_t i = 0; i < num_test_points; ++i)
     {
       double t_val;
       size_t t_row;
@@ -123,7 +123,7 @@ class ValidationRMSETermination
     {
       rmseOld = rmse;
       rmse = 0;
-      for (size_t i = 0; i < num_test_points; i++)
+      for (size_t i = 0; i < num_test_points; ++i)
       {
         size_t t_row = test_points(i, 0);
         size_t t_col = test_points(i, 1);