+++ src/mlpack/methods/reinforcement_learning/q_learning.hpp
@@ -90,10 +90,14 @@ class QLearning
   ~QLearning();
 
   /**
-   * Execute a step in an episode.
-   * @return Reward for the step.
+   * Trains the DQN agent(non-categorical).
    */
-  double Step();
+  void TrainAgent();
+
+  /**
+   * Select an action, given an agent.
+   */
+  void SelectAction();
 
   /**
    * Execute an episode.
@@ -101,9 +105,9 @@ class QLearning
    */
   double Episode();
 
-  /**
-   * @return Total steps from beginning.
-   */
+  //! Modify total steps from beginning.
+  size_t& TotalSteps() { return totalSteps; }
+  //! Get total steps from beginning.
   const size_t& TotalSteps() const { return totalSteps; }
 
   //! Modify the state of the agent.