+++ src/mlpack/methods/reinforcement_learning/q_learning_impl.hpp
@@ -118,40 +118,19 @@ template <
   typename BehaviorPolicyType,
   typename ReplayType
 >
-double QLearning<
+void QLearning<
   EnvironmentType,
   NetworkType,
   UpdaterType,
   BehaviorPolicyType,
   ReplayType
->::Step()
+>::TrainAgent()
 {
-  // Get the action value for each action at current state.
-  arma::colvec actionValue;
-  learningNetwork.Predict(state.Encode(), actionValue);
-
-  // Select an action according to the behavior policy.
-  action = policy.Sample(actionValue, deterministic, config.NoisyQLearning());
-
-  // Interact with the environment to advance to next state.
-  StateType nextState;
-  double reward = environment.Sample(state, action, nextState);
-
-  // Store the transition for replay.
-  replayMethod.Store(state, action, reward,
-      nextState, environment.IsTerminal(nextState));
-
-  // Update current state.
-  state = nextState;
-
-  if (deterministic || totalSteps < config.ExplorationSteps())
-    return reward;
-
   // Start experience replay.
 
   // Sample from previous experience.
   arma::mat sampledStates;
-  arma::icolvec sampledActions;
+  std::vector<ActionType> sampledActions;
   arma::colvec sampledRewards;
   arma::mat sampledNextStates;
   arma::icolvec isTerminal;
@@ -180,6 +159,8 @@ double QLearning<
   arma::mat target;
   learningNetwork.Forward(sampledStates, target);
 
+  double discount = std::pow(config.Discount(), replayMethod.NSteps());
+
   /**
    * If the agent is at a terminal state, then we don't need to add the
    * discounted reward. At terminal state, the agent wont perform any
@@ -187,15 +168,8 @@ double QLearning<
    */
   for (size_t i = 0; i < sampledNextStates.n_cols; ++i)
   {
-    if (isTerminal[i])
-    {
-      target(sampledActions(i), i) = sampledRewards(i);
-    }
-    else
-    {
-      target(sampledActions(i), i) = sampledRewards(i) + config.Discount() *
-          nextActionValues(bestActions(i), i);
-    }
+    target(sampledActions[i].action, i) = sampledRewards(i) + discount *
+        nextActionValues(bestActions(i), i) * (1 - isTerminal[i]);
   }
 
   // Learn from experience.
@@ -216,8 +190,35 @@ double QLearning<
     learningNetwork.ResetNoise();
     targetNetwork.ResetNoise();
   }
+  // Update target network.
+  if (totalSteps % config.TargetNetworkSyncInterval() == 0)
+    targetNetwork = learningNetwork;
 
-  return reward;
+  if (totalSteps > config.ExplorationSteps())
+    policy.Anneal();
+}
+
+template <
+  typename EnvironmentType,
+  typename NetworkType,
+  typename UpdaterType,
+  typename BehaviorPolicyType,
+  typename ReplayType
+>
+void QLearning<
+  EnvironmentType,
+  NetworkType,
+  UpdaterType,
+  BehaviorPolicyType,
+  ReplayType
+>::SelectAction()
+{
+  // Get the action value for each action at current state.
+  arma::colvec actionValue;
+  learningNetwork.Predict(state.Encode(), actionValue);
+
+  // Select an action according to the behavior policy.
+  action = policy.Sample(actionValue, deterministic, config.NoisyQLearning());
 }
 
 template <
@@ -238,34 +239,31 @@ double QLearning<
   // Get the initial state from environment.
   state = environment.InitialSample();
 
-  // Track the steps in this episode.
-  size_t steps = 0;
-
   // Track the return of this episode.
   double totalReturn = 0.0;
 
   // Running until get to the terminal state.
   while (!environment.IsTerminal(state))
   {
-    if (config.StepLimit() && steps >= config.StepLimit())
-      break;
+    SelectAction();
 
-    totalReturn += Step();
-    steps++;
+    // Interact with the environment to advance to next state.
+    StateType nextState;
+    double reward = environment.Sample(state, action, nextState);
 
-    if (deterministic)
-      continue;
-
+    totalReturn += reward;
     totalSteps++;
 
-    // Update target network
-    if (totalSteps % config.TargetNetworkSyncInterval() == 0)
-      targetNetwork = learningNetwork;
+    // Store the transition for replay.
+    replayMethod.Store(state, action, reward, nextState,
+        environment.IsTerminal(nextState), config.Discount());
+    // Update current state.
+    state = nextState;
 
-    if (totalSteps > config.ExplorationSteps())
-      policy.Anneal();
+    if (deterministic || totalSteps < config.ExplorationSteps())
+      continue;
+    TrainAgent();
   }
-
   return totalReturn;
 }